% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gs2020}
\alias{gs2020}
\title{gs2020}
\format{
A dataframe containing 7007 rows and 9 variables  
\itemize{
  \item cycle time 
  \item district_final individual 
  \item general_sharetotal_A_all outcome 
  \item general_sharetotal_B_all other variable 
  \item general_sharetotal_H_all other variable 
  \item cand_A_all treatment
  \item cand_B_all covariate 1
  \item cand_H_all covariate 2
  \item dfe id 
}
}
\source{
Grumbach, Jacob M., and Alexander Sahn. 2020.
 “Race and Representation in Campaign Finance.” 
 American Political Science Review 114 (1): 206–21. https://doi.org/10.1017/S0003055419000637.
}
\usage{
data(paneltools)
}
\description{
data from Grumbach and Sahn (2020)
which uses district-election year level TSCS data from US House general elections between 1980 and 2012
}
