% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hh2019}
\alias{hh2019}
\title{hh2019}
\format{
A dataframe containing 22971 rows and 4 variables 
\itemize{  
  \item bfs id  
  \item year time  
  \item nat_rate_ord outcome  
  \item indirect treatment  
}
}
\source{
Hainmueller, Jens, and Dominik Hangartner. 2019.
 “Does Direct Democracy Hurt Immigrant Minorities? Evidence from Naturalization Decisions in Switzerland.” 
 American Journal of Political Science 63 (3): 530–47.
}
\usage{
data(paneltools)
}
\description{
an empirical example from Hainmueller and Hopkins (2019), 
in which the authors investigate the effects of indirect democracy (versus direct democracy) 
on naturalization rates in Switzerland using municipality-year level panel data from 1991 to 2009.
}
