% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.CIplot.R
\name{matching.CIplot}
\alias{matching.CIplot}
\title{matching.CIplot}
\usage{
matching.CIplot(df)
}
\arguments{
\item{df}{A dataframe containing the periods, estimates, and the CI}
}
\value{
\code{matching.CIplot} return a ggplot object of the event study plot.
}
\description{
\code{matching.CIplot} visualize the dynamic treatment effects of the match episode 
and create an event study plot.
}
\author{
Hongyi Jiang <hyjiang2017@nsd.pku.edu.cn>
}
