% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.plot.R
\name{matching.plot}
\alias{matching.plot}
\title{matching.plot}
\usage{
matching.plot(
  Dwide,
  c,
  a,
  b,
  type = "episode",
  this_episode,
  this_cohort = NULL
)
}
\arguments{
\item{Dwide}{wide panel for treatment}

\item{c}{tuning parameter of carryover}

\item{a}{tuning parameter of pre-treatment}

\item{b}{tuning parameter of post-treatment}

\item{type}{"episode" or "cohort"}

\item{this_episode}{the chosen episode, will be used when type == "episode"}

\item{this_cohort}{the chosen cohort, will be used when type == "cohort"}
}
\value{
\code{matching.plot} return a ggplot object of the chosen episode/cohort.
}
\description{
\code{matching.plot} visualize the match episode/cohort and create a grid plot.
}
\author{
Hongyi Jiang <hyjiang2017@nsd.pku.edu.cn>
}
