% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{matching}
\alias{matching}
\title{matching}
\usage{
matching(data, Y, D, X = NULL, index, c, a, b, type = "episode")
}
\arguments{
\item{data}{A long form panel, can be missing or un-balanced}

\item{Y}{name of outcome}

\item{D}{name of treatment}

\item{X}{name of covariates}

\item{index}{name of id and time}

\item{c}{tuning parameter of carryover}

\item{a}{tuning parameter of pre-treatment}

\item{b}{tuning parameter of post-treatment}

\item{type}{"episode" or "cohort"}
}
\value{
\code{matching} return a list containing the wide panel, true id and unit, 
tuning parameters, and each episode/cohort.
}
\description{
\code{matching} generate the episode or cohort according to the tuning parameter
}
\author{
Hongyi Jiang <hyjiang2017@nsd.pku.edu.cn>
}
