% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paneltools-package.R
\name{paneltools-package}
\alias{paneltools-package}
\title{paneltools}
\description{
A Toolkit for Panel Data Analysis
}
\details{
This package provides a set of commonly used functions for panel data analysis, 
including data preprocessing, obtaining relative period to treatment, event study plot visualization,
match and its estimates as well as inference
}
\author{
Hongyi Jiang hyjiang2017@nsd.pku.edu.cn, Peking University
Ziyi Liu ziyiliu2020@gmail.com, University of Chicago 
Yiqing Xu yiqingxu@stanford.edu, Stanford University 
Zhongyu Yin zhongyuy@stanford.edu, Stanford University
}
\keyword{package}
