% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.CI.R
\name{matching.CI}
\alias{matching.CI}
\title{matching.CI}
\usage{
matching.CI(
  sets,
  alpha = 0.05,
  estimand = "Dynamic",
  method = "Simple DID",
  vce = "jackknife",
  nsim = 200,
  parallel = TRUE,
  cores = 4,
  demean = TRUE,
  imbal.tol = FALSE,
  term.alpha = NULL,
  cv = NULL
)
}
\arguments{
\item{sets}{episode generated by match.R}

\item{alpha}{significance level}

\item{estimand}{"ATT" or "Dynamic"}

\item{method}{"Simple DID" now, can be added in later version}

\item{vce}{"jackknife" or "bootstrap"}

\item{nsim}{numbers of bootstrap, also used for limit the number of jackknife}

\item{parallel}{whether do parallel computing}

\item{cores}{cores used in the parallel computing}

\item{demean}{used in mbal/kbal}

\item{imbal.tol}{used in mbal/kbal, if FALSE, will only keep the episodes with bias.ratio < 1e-5}

\item{term.alpha}{named vector of ridge penalties, only takes 0 or 1}

\item{cv}{whether to use cross validation}
}
\value{
\code{matching.CI} return a dataframe containing the periods, estimates, and the CI
}
\description{
\code{matching.CI} give the estimates of the match episode, using different 
estimator or inference.
}
\author{
Hongyi Jiang <hyjiang2017@nsd.pku.edu.cn>
}
