\name{paneltools-package}
\alias{paneltools-package}
\docType{package}
\title{Paneltools}
\description{A Toolkit for Panel Data Analysis}
\details{
This package provides a set of commonly used functions for panel data analysis, including data preprocessing, obtaining relative period to treatment, and event study plot visualization.
}
\author{
Ziyi Liu ziyiliu2020@gmail.com, University of Chicago 

Yiqing Xu yiqingxu@stanford.edu, Stanford University 

Zhongyu Yin zhongyuy@stanford.edu, Stanford University
}

